Maps/Warnings: Reduce the amount of warnings/err...

+++ src/tools/ToolsCata/vmap_tools/vmap4_extractor/wmo.cpp
@@ -394,7 +394,7 @@ int WMOGroup::ConvertToVMAPGroupWmo(FILE *output, WMORoot *rootWMO, bool precise
         fwrite(VERT, 4, 3, output);
         for (uint32 i = 0; i<nVertices; ++i)
             if (IndexRenum[i] >= 0)
-                check -= fwrite(MOVT + 3 * i, sizeof(float), 3, output);
+                check -= static_cast<int>(fwrite(MOVT+3*i, sizeof(float), 3, output));
 
         assert(check == 0);
